Changed request to GetHostsGetHostListAsync, works
